Fix accept-invite Suspense build error
